Split report tests out
