+++ test/test_helpers.js
@@ -435,295 +435,6 @@ describe('helpers.js', function() {
         });
     });
 
-    describe('cleanAnalyzeDataEmptyProps', () => {
-        const contractJSON = `${__dirname}/sample-truffle/simple_dao/build/contracts/SimpleDAO.json`;
-        let truffleJSON;
-
-        beforeEach(done => {
-            fs.readFile(contractJSON, 'utf8', (err, data) => {
-                if (err) return done(err);
-                truffleJSON = JSON.parse(data);
-                done();
-            });
-        });
-
-        it('should return complete input data when all fields are present', () => {
-            const stub = sinon.stub();
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(!stub.called);
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit bytecode when bytecode is empty', () => {
-            const stub = sinon.stub();
-            truffleJSON.bytecode = '';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.bytecode;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit bytecode when bytecode is 0x', () => {
-            const stub = sinon.stub();
-            truffleJSON.bytecode = '0x';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.bytecode;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit deployedBytecode when deployedBytecode is empty', () => {
-            const stub = sinon.stub();
-            truffleJSON.deployedBytecode = '';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.deployedBytecode;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit deployedBytecode when deployedBytecode is 0x', () => {
-            const stub = sinon.stub();
-            truffleJSON.deployedBytecode = '0x';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.deployedBytecode;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit sourceMap when sourceMap is empty', () => {
-            const stub = sinon.stub();
-            truffleJSON.sourceMap = '';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.sourceMap;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit deployedSourceMap when deployedSourceMap is empty', () => {
-            const stub = sinon.stub();
-            truffleJSON.deployedSourceMap = '';
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, true, stub);
-            assert.ok(stub.called);
-            delete truffleJSON.deployedSourceMap;
-            assert.deepEqual(result, truffleJSON);
-        });
-
-        it('should omit empty fields but not log  when debug is false', () => {
-            const stub = sinon.stub();
-            truffleJSON.deployedSourceMap = '';
-            truffleJSON.sourceMap = null;
-            truffleJSON.bytecode = '0x';
-            delete truffleJSON.deployedBytecode;
-            const result = rewiredHelpers.cleanAnalyzeDataEmptyProps(truffleJSON, false, stub);
-            delete truffleJSON.sourceMap;
-            delete truffleJSON.deployedSourceMap;
-            delete truffleJSON.bytecode;
-            delete truffleJSON.deployedBytecode;
-            assert.ok(!stub.called);
-            assert.deepEqual(result, truffleJSON);
-        });
-    });
-
-    describe('doReport', () => {
-        let loggerStub;
-        let errorStub;
-        let config;
-
-        beforeEach(() => {
-            loggerStub = sinon.stub();
-            errorStub = sinon.stub();
-
-            config = {
-                logger: {
-                    log: loggerStub,
-                    error: errorStub,
-                },
-                json: true,
-            };
-        });
-
-        it('should return 0 when no errors, no issues, and no logs', async () => {
-            const results = {
-                "errors": [],
-                "objects": [
-                    {
-                        issues: [
-                            {
-                                "issues": [],
-                            }
-                        ],
-                        logs: []
-                    },
-                    {
-                        issues: [
-                            {
-                                "issues": []
-                            }
-                        ],
-                        logs: []
-                    }
-                ]
-            };
-            const ret = rewiredHelpers.__get__('doReport')(config, results.objects, results.errors);
-            assert.ok(!loggerStub.calledWith('MythX Logs:'.yellow));
-            assert.ok(!errorStub.calledWith('Internal MythX errors encountered:'.red));
-            assert.equal(ret, 0);
-        });
-
-        it('should return 1 when errors is 1 or more', async () => {
-            const results = {
-                "errors": [
-                    {
-                      "status": "Error"
-                    }
-                ],
-                "objects": [
-                    {
-                        issues: [
-                            {
-                                "issues": [],
-                            }
-                        ],
-                        logs: []
-                    },
-                    {
-                        issues: [
-                            {
-                                "issues": []
-                            }
-                        ],
-                        logs: []
-                    }
-                ]
-            };
-            const ret = rewiredHelpers.__get__('doReport')(config, results.objects, results.errors);
-            assert.ok(!loggerStub.calledWith('MythX Logs:'.yellow));
-            assert.ok(errorStub.calledWith('Internal MythX errors encountered:'.red));
-            assert.equal(ret, 1);
-        });
-
-        it('should return 1 when issues is 1 or more', () => {
-            const results = {
-                "errors": [],
-                "objects": [
-                    {
-                        issues: [
-                            {
-                                "issues": [{
-                                    'description': {
-                                        'head': 'Head message',
-                                        'tail': 'Tail message'
-                                    },
-                                    'locations': [{
-                                        'sourceMap': '444:1:0'
-                                    }],
-                                    'severity': 'High',
-                                    'swcID': 'SWC-000',
-                                    'swcTitle': 'Test Title'
-                                }],
-                            }
-                        ],
-                        logs: []
-                    },
-                    {
-                        issues: [
-                            {
-                                "issues": []
-                            }
-                        ],
-                        logs: []
-                    }
-                ]
-            };
-            const ret = rewiredHelpers.__get__('doReport')(config, results.objects, results.errors);
-            assert.ok(!loggerStub.calledWith('MythX Logs:'.yellow));
-            assert.ok(!errorStub.calledWith('Internal MythX errors encountered:'.red));
-            assert.equal(ret, 1);
-        });
-
-        it('should return 0 when logs is 1 or more with debug', async () => {
-            config.debug = true;
-            const results = {
-                errors: [],
-                objects: [
-                    {
-                        issues: [
-                            {
-                                "issues": [],
-                            }
-                        ],
-                        logs: [
-                            {
-                                level : 'info',
-                                msg: 'message1',
-                            }
-                        ]
-                    },
-                    {
-                        issues: [
-                            {
-                                "issues": [],
-                            }
-                        ],
-                        logs: []
-                    }
-                ]
-            };
-            const ret = rewiredHelpers.__get__('doReport')(config, results.objects, results.errors);
-            assert.ok(loggerStub.calledWith('MythX Logs:'.yellow));
-            assert.ok(!errorStub.calledWith('Internal MythX errors encountered:'.red));
-            assert.equal(ret, 1);
-        });
-    });
-
-    describe('ghettoReport', () => {
-        let loggerStub = sinon.stub();
-        beforeEach(() => {
-            loggerStub = sinon.stub();
-        });
-
-        it('should return 0 when issues count is 0', () => {
-            const results = [{
-                "issues": [],
-            }];
-            const ret = rewiredHelpers.__get__('ghettoReport')(loggerStub, results);
-            assert.ok(loggerStub.calledWith('No issues found'));
-            assert.equal(ret, 0);
-        });
-
-        it('should return 1 when issues count is 1 or more', () => {
-            const results = [{
-                'sourceFormat': 'evm-byzantium-bytecode',
-                'sourceList': [
-                    'list1', 'list2'
-                ],
-                'sourceType': 'raw-bytecode',
-                'issues': [{
-                    'description': {
-                        'head': 'Head message',
-                        'tail': 'Tail message'
-                    },
-                    'locations': [{
-                        'sourceMap': '444:1:0'
-                    }],
-                    'severity': 'High',
-                    'swcID': 'SWC-000',
-                    'swcTitle': 'Test Title'
-                }],
-                'meta': {
-                    'selected_compiler': '0.5.0',
-                    'error': [],
-                    'warning': []
-                }
-            }];
-
-            const ret = rewiredHelpers.__get__('ghettoReport')(loggerStub, results);
-            assert.ok(!loggerStub.calledWith('No issues found'));
-            assert.ok(loggerStub.calledWith('list1, list2'.underline));
-            assert.ok(loggerStub.calledWith(yaml.safeDump(results[0].issues[0], {'skipInvalid': true})));
-            assert.equal(ret, 1);
-        });
-    });
 
     describe('prepareConfig', () => {
         it('should return a numeric severityThreshold', () => {
